remove: deal page view setting 버튼 삭제
