+++ src/apis/deal/useDeal.ts
@@ -80,28 +80,3 @@ export const postDealColumnCreate = async ({
     throw error;
   }
 };
-
-export const patchDealCell = async ({
-  columnId,
-  rowId,
-  value,
-  peopleId,
-}: {
-  columnId: number;
-  rowId: number;
-  value: string;
-  peopleId: number;
-}) => {
-  try {
-    const response = await axiosClient.patch(`/workspaces/cells`, {
-      columnId,
-      rowId,
-      value,
-      peopleId,
-    });
-
-    return response.data;
-  } catch (error) {
-    console.error(error);
-  }
-};
