+++ src/components/common/modal/deals/DealCompanyEditor.tsx
@@ -1,0 +1,29 @@
+// src/components/DealCompanyEditor.tsx
+import React, { forwardRef, useImperativeHandle, useState } from 'react';
+import { Select } from 'antd';
+
+const { Option } = Select;
+
+const DealCompanyEditor = forwardRef((props: any, ref) => {
+  const [value, setValue] = useState(props.value);
+
+  useImperativeHandle(ref, () => ({
+    getValue: () => value,
+  }));
+
+  const handleChange = (value: string) => {
+    setValue(value);
+  };
+
+  return (
+    <Select value={value} onChange={handleChange} style={{ width: '100%' }}>
+      {props.companies.map((company: any) => (
+        <Option key={company.companyId} value={company.companyId}>
+          {company.companyName}
+        </Option>
+      ))}
+    </Select>
+  );
+});
+
+export default DealCompanyEditor;
